refactor(cli): split install root orchestration
